+++ pipen/_job_caching.py
@@ -1,21 +1,23 @@
 """Provide JobCaching class that implements caching for jobs"""
+
 from __future__ import annotations
+from typing import TYPE_CHECKING
 
-from pathlib import Path
-
-from cloudpathlib import AnyPath
 from diot import Diot
 from simpleconf import Config
 
 from .defaults import ProcInputType, ProcOutputType
-from .utils import get_mtime, path_rmtree, path_is_symlink
+from .utils import get_mtime, path_is_symlink
+
+if TYPE_CHECKING:
+    from xqute.path import DualPath
 
 
 class JobCaching:
     """Provide caching functionality of jobs"""
 
     @property
-    def signature_file(self) -> Path:
+    def signature_file(self) -> DualPath:
         """Get the path to the signature file
 
         Returns:
@@ -36,33 +38,47 @@ class JobCaching:
         except Exception:  # pragma: no cover
             max_mtime = 0
 
+        # Make self.input serializable
+        input_data = {}
         for inkey, intype in self.proc.input.type.items():
             if intype == ProcInputType.VAR:
+                input_data[inkey] = self.input[inkey]
                 continue
 
-            if (
-                intype in (ProcInputType.FILE, ProcInputType.DIR)
-                and self.input[inkey] is not None
-            ):
-                max_mtime = max(max_mtime, get_mtime(self.input[inkey], dirsig))
+            if intype in (ProcInputType.FILE, ProcInputType.DIR):
+                if self.input[inkey] is None:
+                    input_data[inkey] = None
+                else:
+                    input_data[inkey] = str(self.input[inkey].spec)
+                    max_mtime = max(
+                        max_mtime,
+                        get_mtime(self.input[inkey].spec, dirsig),
+                    )
 
-            if (
-                intype in (ProcInputType.FILES, ProcInputType.DIRS)
-                and self.input[inkey] is not None
-            ):
-                for file in self.input[inkey]:
-                    max_mtime = max(max_mtime, get_mtime(file, dirsig))
+            if intype in (ProcInputType.FILES, ProcInputType.DIRS):
+                if self.input[inkey] is None:  # pragma: no cover
+                    input_data[inkey] = None
+                else:
+                    input_data[inkey] = []
+                    for file in self.input[inkey]:
+                        input_data[inkey].append(str(file.spec))
+                        max_mtime = max(max_mtime, get_mtime(file.spec, dirsig))
 
+        # Make self.output serializable
+        output_data = {}
         for outkey, outval in self._output_types.items():
             if outval in (ProcOutputType.FILE, ProcInputType.DIR):
-                max_mtime = max(max_mtime, get_mtime(self.output[outkey], dirsig))
+                output_data[outkey] = str(self.output[outkey].spec)
+                max_mtime = max(max_mtime, get_mtime(self.output[outkey].spec, dirsig))
+            else:
+                output_data[outkey] = self.output[outkey]
 
         signature = {
             "input": {
                 "type": self.proc.input.type,
-                "data": self.input,
+                "data": input_data,
             },
-            "output": {"type": self._output_types, "data": self.output},
+            "output": {"type": self._output_types, "data": output_data},
             "ctime": float("inf") if max_mtime == 0 else max_mtime,
         }
         with self.signature_file.open("w") as f:
@@ -75,14 +91,14 @@ class JobCaching:
             if outval not in (ProcOutputType.FILE, ProcOutputType.DIR):
                 continue
 
-            path = AnyPath(self.output[outkey])
+            path = self.output[outkey].spec
             if not path.exists() and path_is_symlink(path):  # dead link
                 path.unlink()
             elif path.exists():
                 if not path.is_dir():
                     path.unlink()
                 else:
-                    path_rmtree(path)
+                    path.rmtree(ignore_errors=True)
                     path.mkdir()
 
     async def _check_cached(self) -> bool:
@@ -104,11 +120,9 @@ class JobCaching:
             # check if inputs/outputs are still the same
             if (
                 signature.input.type != self.proc.input.type
-                or signature.input.data != self.input
                 or signature.output.type != self._output_types
-                or signature.output.data != self.output
             ):
-                self.log("debug", "Not cached (input or output is different)")
+                self.log("debug", "Not cached (input or output types are different)")
                 return False
 
             # check if any script file is newer
@@ -122,13 +136,47 @@ class JobCaching:
                 )
                 return False
 
+            # Check if input is different
             for inkey, intype in self.proc.input.type.items():
+                sig_indata = signature.input.data.get(inkey)
 
-                if intype == ProcInputType.VAR or self.input[inkey] is None:
-                    continue  # pragma: no cover, covered, a bug of pytest-cov
+                if intype == ProcInputType.VAR:
+                    if sig_indata != self.input[inkey]:
+                        self.log(
+                            "debug",
+                            "Not cached (input %s:%s is different)",
+                            inkey,
+                            intype,
+                        )
+                        return False
 
-                if intype in (ProcInputType.FILE, ProcInputType.DIR):
-                    if get_mtime(self.input[inkey], dirsig) > signature.ctime + 1e-3:
+                elif int(self.input[inkey] is None) + int(sig_indata is None) == 1:
+                    # one is None, the other is not
+                    self.log(
+                        "debug",
+                        "Not cached (input %s:%s is different; "
+                        "it is <%s> in signature, but <%s> in data)",
+                        inkey,
+                        intype,
+                        type(sig_indata).__name__,
+                        type(self.input[inkey]).__name__,
+                    )
+                    return False
+
+                elif intype in (ProcInputType.FILE, ProcInputType.DIR):
+                    if sig_indata != str(self.input[inkey].spec):
+                        self.log(
+                            "debug",
+                            "Not cached (input %s:%s is different)",
+                            inkey,
+                            intype,
+                        )
+                        return False
+
+                    if (
+                        get_mtime(self.input[inkey].spec, dirsig)
+                        > signature.ctime + 1e-3
+                    ):
                         self.log(
                             "debug",
                             "Not cached (Input file is newer: %s)",
@@ -136,33 +184,92 @@ class JobCaching:
                         )
                         return False
 
-                if intype in (ProcInputType.FILES, ProcInputType.DIRS):
-                    for file in self.input[inkey]:
-                        if get_mtime(file, dirsig) > signature.ctime + 1e-3:
-                            self.log(
-                                "debug",
-                                "Not cached (One of the input files "
-                                "is newer: %s)",
-                                inkey,
-                            )
-                            return False
+                # FILES/DIRS
 
-            for outkey, outval in self._output_types.items():
-                if outval not in (ProcOutputType.FILE, ProcOutputType.DIR):
-                    continue
+                # self.input[inkey] can't be None with intype files/dirs
+                # elif sig_indata is None:  # both None
+                #     continue
 
-                if not AnyPath(self.output[outkey]).exists():
+                elif not isinstance(sig_indata, list):  # pragma: no cover
                     self.log(
                         "debug",
-                        "Not cached (Output file removed: %s)",
-                        outkey,
+                        "Not cached (input %s:%s is different, "
+                        "%s detected in signature)",
+                        inkey,
+                        intype,
+                        type(sig_indata).__name__,
                     )
                     return False
 
-        except (AttributeError, FileNotFoundError):  # pragma: no cover
+                else:  # both list
+                    if len(sig_indata) != len(self.input[inkey]):  # pragma: no cover
+                        self.log(
+                            "debug",
+                            "Not cached (input %s:%s length is different)",
+                            inkey,
+                            intype,
+                        )
+                        return False
+
+                    for i, file in enumerate(self.input[inkey]):
+                        if sig_indata[i] != str(file.spec):  # pragma: no cover
+                            self.log(
+                                "debug",
+                                "Not cached (input %s:%s at index %s is different)",
+                                inkey,
+                                intype,
+                                i,
+                            )
+                            return False
+
+                        if get_mtime(file.spec, dirsig) > signature.ctime + 1e-3:
+                            self.log(
+                                "debug",
+                                "Not cached (input %s:%s at index %s is newer)",
+                                inkey,
+                                intype,
+                                i,
+                            )
+                            return False
+
+            # Check if output is different
+            for outkey, outtype in self._output_types.items():
+                sig_outdata = signature.output.data.get(outkey)
+                if outtype == ProcOutputType.VAR:
+                    if sig_outdata != self.output[outkey]:  # pragma: no cover
+                        self.log(
+                            "debug",
+                            "Not cached (output %s:%s is different)",
+                            outkey,
+                            outtype,
+                        )
+                        return False
+
+                else:  # FILE/DIR
+                    if sig_outdata != str(self.output[outkey].spec):  # pragma: no cover
+                        self.log(
+                            "debug",
+                            "Not cached (output %s:%s is different)",
+                            outkey,
+                            outtype,
+                        )
+                        return False
+
+                    if not self.output[outkey].spec.exists():
+                        self.log(
+                            "debug",
+                            "Not cached (output %s:%s was removed)",
+                            outkey,
+                            outtype,
+                        )
+                        return False
+
+        except Exception as exc:  # pragma: no cover
             # meaning signature is incomplete
             # or any file is deleted
+            self.log("debug", "Not cached (%s)", exc)
             return False
+
         return True
 
     @property
@@ -193,7 +300,8 @@ class JobCaching:
         elif proc_cache == "force":
             try:
                 await self.cache()
-            except FileNotFoundError:  # pragma: no cover
+            except Exception:  # pragma: no cover
+                # FileNotFoundError, google.api_core.exceptions.NotFound, etc
                 out = False
             else:
                 out = True