+++ pipen/utils.py
@@ -1,4 +1,5 @@
 """Provide some utilities"""
+
 from __future__ import annotations
 
 import re
@@ -30,7 +31,7 @@ from typing import (
 
 import diot
 import simplug
-from cloudpathlib import AnyPath, CloudPath
+from yunpath import AnyPath, CloudPath
 from rich.console import Console
 from rich.logging import RichHandler as _RichHandler
 from rich.table import Table
@@ -51,6 +52,7 @@ if TYPE_CHECKING:  # pragma: no cover
     import pandas
     from rich.segment import Segment
     from rich.console import RenderableType
+    from xqute.path import DualPath
 
     from .pipen import Pipen
     from .proc import Proc
@@ -417,7 +419,10 @@ def pipen_banner() -> RenderableType:
     return table
 
 
-def get_mtime(path: str | PathLike | CloudPath, dir_depth: int = 1) -> float:
+def get_mtime(
+    path: str | PathLike | CloudPath | DualPath,
+    dir_depth: int = 1,
+) -> float:
     """Get the modification time of a path.
     If path is a directory, try to get the last modification time of the
     contents in the directory at given dir_depth
@@ -427,6 +432,8 @@ def get_mtime(path: str | PathLike | CloudPath, dir_depth: int = 1) -> float:
     Returns:
         The last modification time of path
     """
+    path = getattr(path, "path", path)
+
     mtime = 0.0
     path = AnyPath(path)
     if not path.exists():
@@ -436,7 +443,7 @@ def get_mtime(path: str | PathLike | CloudPath, dir_depth: int = 1) -> float:
         if dir_depth == 0 or not path.is_dir():
             return path.stat().st_mtime
 
-        for file in path.glob("*"):
+        for file in path.iterdir():
             mtime = max(mtime, get_mtime(file, dir_depth - 1))
 
         return mtime
@@ -446,7 +453,7 @@ def get_mtime(path: str | PathLike | CloudPath, dir_depth: int = 1) -> float:
         if dir_depth == 0 or not path.is_dir():
             return path.lstat().st_mtime
 
-        for file in path.glob("*"):
+        for file in path.iterdir():
             mtime = max(mtime, get_mtime(file, dir_depth - 1))
 
         return mtime
@@ -460,7 +467,7 @@ def get_mtime(path: str | PathLike | CloudPath, dir_depth: int = 1) -> float:
         except Exception:
             return path.stat().st_mtime
 
-    for file in dpath.glob("*"):
+    for file in dpath.iterdir():
         mtime = max(mtime, get_mtime(file, dir_depth - 1))
     return mtime
 
@@ -483,9 +490,7 @@ def is_subclass(obj: Any, cls: type) -> bool:
         return False
 
 
-def load_entrypoints(
-    group: str
-) -> Iterable[Tuple[str, Any]]:  # pragma: no cover
+def load_entrypoints(group: str) -> Iterable[Tuple[str, Any]]:  # pragma: no cover
     """Load objects from setuptools entrypoints by given group name
 
     Args:
@@ -576,6 +581,7 @@ def mark(**kwargs) -> Callable[[type], type]:
     Returns:
         The decorator
     """
+
     def decorator(cls: type) -> type:
         if not getattr(cls, "__meta__", None):
             cls.__meta__ = {}
@@ -648,7 +654,7 @@ def _get_obj_from_spec(spec: str) -> Any:
 
 
 async def load_pipeline(
-    obj: str | Type[Proc] | Type[ProcGroup] | Type[Pipen],
+    obj: str | Type[Proc] | Type[ProcGroup] | Type[Pipen] | Pipen,
     argv0: str | None = None,
     argv1p: Sequence[str] | None = None,
     **kwargs: Any,
@@ -717,23 +723,23 @@ async def load_pipeline(
 
         elif isinstance(obj, type) and issubclass(obj, Pipen):
             # Avoid "pipeline" to be used as pipeline name by varname
-            (pipeline, ) = (obj(**kwargs), )  # type: ignore
+            (pipeline,) = (obj(**kwargs),)  # type: ignore
 
         elif isinstance(obj, Pipen):
-            pipeline._kwargs.update(kwargs)
+            pipeline._kwargs.update(kwargs)  # type: ignore
 
         # Initialize the pipeline so that the arguments definied by
         # other plugins (i.e. pipen-args) to take in place.
-        pipeline.workdir = Path(pipeline.config.workdir).joinpath(
+        pipeline.workdir = Path(pipeline.config.workdir).joinpath(  # type: ignore
             kwargs.get("name", pipeline.name)
         )
-        await pipeline._init()
-        pipeline.workdir.mkdir(parents=True, exist_ok=True)
-        pipeline.build_proc_relationships()
+        await pipeline._init()  # type: ignore
+        pipeline.workdir.mkdir(parents=True, exist_ok=True)  # type: ignore
+        pipeline.build_proc_relationships()  # type: ignore
     finally:
         sys.argv = old_argv
 
-    return pipeline
+    return pipeline  # type: ignore
 
 
 def is_loading_pipeline(*flags: str, argv: Sequence[str] | None = None) -> bool:
@@ -768,7 +774,7 @@ def is_loading_pipeline(*flags: str, argv: Sequence[str] | None = None) -> bool:
     return False  # pragma: no cover
 
 
-def path_is_symlink(path: Path | CloudPath) -> bool:
+def path_is_symlink(path: Path | CloudPath | DualPath) -> bool:
     """Check if a path is a symlink.
 
     CloudPath.is_symlink() is not implemented yet, so we need to check
@@ -780,6 +786,7 @@ def path_is_symlink(path: Path | CloudPath) -> bool:
     Returns:
         True if the path is a symlink, otherwise False
     """
+    path = getattr(path, "path", path)
     if isinstance(path, Path):
         if path.is_symlink():
             return True
@@ -795,8 +802,8 @@ def path_is_symlink(path: Path | CloudPath) -> bool:
 
 
 def path_symlink_to(
-    src: Path | CloudPath,
-    dst: Path | CloudPath,
+    src: Path | CloudPath | DualPath,
+    dst: Path | CloudPath | DualPath,
     target_is_directory: bool = False,
 ) -> None:
     """Create a symbolic link pointing to src named dst.
@@ -806,20 +813,10 @@ def path_symlink_to(
         dst: The destination path
         target_is_directory: If True, the symbolic link will be to a directory.
     """
+    src = getattr(src, "path", src)
+    dst = getattr(dst, "path", dst)
     if isinstance(dst, CloudPath) or isinstance(src, CloudPath):
         # Create a fake symlink file for cloud paths
         src.write_text(f"symlink:{dst}")
     else:
         src.symlink_to(dst, target_is_directory=target_is_directory)
-
-
-def path_rmtree(path: Path | CloudPath) -> None:
-    """Remove a directory tree for both local and cloud paths.
-
-    Args:
-        path: The path to remove
-    """
-    if isinstance(path, CloudPath):
-        path.rmtree()
-    else:
-        shutil.rmtree(path)