+++ tests/test_scheduler.py
@@ -1,4 +1,6 @@
 import pytest
+from unittest.mock import MagicMock
+
 
 from pipen.scheduler import (
     get_scheduler,
@@ -6,6 +8,7 @@ from pipen.scheduler import (
     SgeScheduler,
     SshScheduler,
     SlurmScheduler,
+    GbatchScheduler,
     NoSuchSchedulerError,
 )
 
@@ -36,5 +39,45 @@ def test_get_scheduler():
     ssh = get_scheduler(ssh)
     assert ssh is SshScheduler
 
+    gbatch = get_scheduler("gbatch")
+    assert gbatch is GbatchScheduler
+
+    gbatch = get_scheduler(gbatch)
+    assert gbatch is GbatchScheduler
+
     with pytest.raises(NoSuchSchedulerError):
         get_scheduler("nosuchscheduler")
+
+
+def test_gbatch_scheduler_post_init():
+    gbatch = get_scheduler("gbatch")(
+        project="test_project",
+        location="test_location",
+        workdir="gs://test-bucket/workdir",
+    )
+    pipeline_outdir = MagicMock(_no_prefix="a/b/c")
+    pipeline = MagicMock(outdir=pipeline_outdir)
+    proc = MagicMock(pipeline=pipeline)
+    proc.name = "test_proc"
+    gbatch.post_init(proc)
+
+    assert str(gbatch.workdir.path) == "gs://test-bucket/workdir"
+    assert (
+        str(gbatch.workdir.mounted) == f"{GbatchScheduler.MOUNTED_METADIR}/{proc.name}"
+    )
+    assert (
+        gbatch.config.taskGroups[0].taskSpec.volumes[-1].mountPath
+        == f"{GbatchScheduler.MOUNTED_OUTDIR}"
+    )
+    assert (
+        gbatch.config.taskGroups[0].taskSpec.volumes[-1].gcs.remotePath
+        == "a/b/c"
+    )
+    assert (
+        gbatch.config.taskGroups[0].taskSpec.volumes[-2].mountPath
+        == f"{GbatchScheduler.MOUNTED_METADIR}/{proc.name}"
+    )
+    assert (
+        gbatch.config.taskGroups[0].taskSpec.volumes[-2].gcs.remotePath
+        == "test-bucket/workdir"
+    )